Move `formInput` styles to `utils/styles`
