+++ www/src/components/hubspot-form.js
@@ -3,7 +3,7 @@ import HubspotForm from "react-hubspot-form"
 import presets, { colors } from "../utils/presets"
 import { rhythm, options } from "../utils/typography"
 import hex2rgba from "hex2rgba"
-import { formInput } from "../utils/form-styles"
+import { formInput } from "../utils/styles"
 import { buttonStyles } from "../utils/styles"
 
 export default class GatsbyHubspotForm extends Component {
